requirements.txt
Flask
python-docx
google-generativeai
streamlit
fpdf
markdown
beautifulsoup4